Progress Towards Event Deletion
Struggling to create an onClickListener for the remove button either in the Fragment or in the Adapter. In the List Fragment, there is no way to access the button attributes of the ViewHolder to set the listener. In the Adapter, there is no way to call the updateRecyclerView() function from the ListFragment so that the event gets removed from the screen.