Fixed event editing bug and made settings page safer
There was a bug because editing or deleting an event did not originally delete it from the database, so reloading the app would cause the old event to still appear. Additionally, I made the settings change to defaults when signing in if the data takes too long to load so it doesn't just crash.