Added old event deletion
When the user is logged in it checks if the events were last cleared on a previous day, so in essence every time the day changes all events will get deleted (This is a DAY planner after all). This fixes the issue where events without a start time would still show up weeks later, and clears unnecessary data from the database. Firestore's TTL could not be used because it can take up to 72 hours to delete data. Also removed some commented test data and started on adding support for recurring events.